requirements.txt
beautifulsoup4
langchain
langgraph
langchainhub
langchain-community
tavily-python
langchain-chroma
python-dotenv
pytest
langchain-openai
